requirements.txt
anthropic
flask
python-telegram-bot
pymupdf
python-docx
